+++ nodejs/configManager.js
@@ -128,7 +128,23 @@ class ConfigManager {
             });
         }
 
-        return this.bigIp.deviceInfo()
+        let provisionedModules = [];
+        return Promise.resolve()
+            .then(() => {
+                return this.bigIp.list('/tm/sys/provision');
+            })
+            .then((provisioning) => {
+                provisionedModules = provisioning
+                    .filter((module) => {
+                        return module.level !== 'none';
+                    })
+                    .map((module) => {
+                        return module.name;
+                    });
+            })
+            .then(() => {
+                return this.bigIp.deviceInfo();
+            })
             .then((deviceInfo) => {
                 this.configId = deviceInfo.machineId;
                 return getTokenMap.call(this, deviceInfo);
@@ -139,26 +155,33 @@ class ConfigManager {
 
                 // get a list of iControl Rest queries asking for the config items and selecting the
                 // properties we want
-                this.configItems.forEach((configItem) => {
-                    const query = { $filter: 'partition eq Common' };
-                    const selectProperties = getPropertiesOfInterest(configItem.properties);
-                    if (selectProperties.length > 0) {
-                        query.$select = selectProperties.join(',');
-                    }
-                    const encodedQuery = querystring.stringify(query);
-                    const options = {};
-                    let path = `${configItem.path}?${encodedQuery}`;
+                this.configItems
+                    .filter((configItem) => {
+                        if (!configItem.requiredModule) {
+                            return true;
+                        }
+                        return provisionedModules.indexOf(configItem.requiredModule) > -1;
+                    })
+                    .forEach((configItem) => {
+                        const query = { $filter: 'partition eq Common' };
+                        const selectProperties = getPropertiesOfInterest(configItem.properties);
+                        if (selectProperties.length > 0) {
+                            query.$select = selectProperties.join(',');
+                        }
+                        const encodedQuery = querystring.stringify(query);
+                        const options = {};
+                        let path = `${configItem.path}?${encodedQuery}`;
 
-                    // do any replacements
-                    path = path.replace(hostNameRegex, tokenMap.hostName);
-                    path = path.replace(deviceNameRegex, tokenMap.deviceName);
+                        // do any replacements
+                        path = path.replace(hostNameRegex, tokenMap.hostName);
+                        path = path.replace(deviceNameRegex, tokenMap.deviceName);
 
-                    if (configItem.silent) {
-                        options.silent = configItem.silent;
-                    }
+                        if (configItem.silent) {
+                            options.silent = configItem.silent;
+                        }
 
-                    promises.push(this.bigIp.list(path, null, cloudUtil.SHORT_RETRY, options));
-                });
+                        promises.push(this.bigIp.list(path, null, cloudUtil.SHORT_RETRY, options));
+                    });
 
                 return Promise.all(promises);
             })