+++ test/nodejs/configManagerTests.js
@@ -19,22 +19,18 @@
 const assert = require('assert');
 const URL = require('url');
 
-/* eslint-disable global-require, quote-props, quotes */
+const ConfigManager = require('../../nodejs/configManager');
 
 describe('configManager', () => {
     const hostname = 'myhost.bigip.com';
     const deviceName = 'device1';
 
     let listResponses;
-    let configItems;
-    let ConfigManager;
     let bigIpMock;
     let state;
     let doState;
 
     before(() => {
-        ConfigManager = require('../../nodejs/configManager');
-
         bigIpMock = {
             deviceInfo() {
                 return Promise.resolve({ hostname });
@@ -55,7 +51,8 @@ describe('configManager', () => {
                     hostname,
                     name: deviceName
                 }
-            ]
+            ],
+            '/tm/sys/provision': []
         };
         state = {};
         doState = {
@@ -66,11 +63,11 @@ describe('configManager', () => {
 
     it('should handle simple string values', () => {
         return new Promise((resolve, reject) => {
-            configItems = [
+            const configItems = [
                 {
-                    "path": "/tm/sys/global-settings",
-                    "properties": [
-                        { "id": "hostname" }
+                    path: '/tm/sys/global-settings',
+                    properties: [
+                        { id: 'hostname' }
                     ]
                 }
             ];
@@ -94,13 +91,13 @@ describe('configManager', () => {
 
     it('should handle objects', () => {
         return new Promise((resolve, reject) => {
-            configItems = [
+            const configItems = [
                 {
-                    "path": "/tm/sys/ntp",
-                    "schemaClass": "NTP",
-                    "properties": [
-                        { "id": "servers" },
-                        { "id": "timezone" }
+                    path: '/tm/sys/ntp',
+                    schemaClass: 'NTP',
+                    properties: [
+                        { id: 'servers' },
+                        { id: 'timezone' }
                     ]
                 }
             ];
@@ -124,14 +121,14 @@ describe('configManager', () => {
 
     it('should handle arrays', () => {
         return new Promise((resolve, reject) => {
-            configItems = [
+            const configItems = [
                 {
-                    "path": "/tm/net/route",
-                    "schemaClass": "Route",
-                    "properties": [
-                        { "id": "gw" },
-                        { "id": "network" },
-                        { "id": "mtu" }
+                    path: '/tm/net/route',
+                    schemaClass: 'Route',
+                    properties: [
+                        { id: 'gw' },
+                        { id: 'network' },
+                        { id: 'mtu' }
                     ]
                 }
             ];
@@ -172,14 +169,14 @@ describe('configManager', () => {
 
     it('should handle arrays when none are already defined on BIG-IP < 14.x', () => {
         return new Promise((resolve, reject) => {
-            configItems = [
+            const configItems = [
                 {
-                    "path": "/tm/net/route",
-                    "schemaClass": "Route",
-                    "properties": [
-                        { "id": "gw" },
-                        { "id": "network" },
-                        { "id": "mtu" }
+                    path: '/tm/net/route',
+                    schemaClass: 'Route',
+                    properties: [
+                        { id: 'gw' },
+                        { id: 'network' },
+                        { id: 'mtu' }
                     ]
                 }
             ];
@@ -200,14 +197,14 @@ describe('configManager', () => {
 
     it('should handle arrays when none are already defined on BIG-IP > 14.x', () => {
         return new Promise((resolve, reject) => {
-            configItems = [
+            const configItems = [
                 {
-                    "path": "/tm/net/route",
-                    "schemaClass": "Route",
-                    "properties": [
-                        { "id": "gw" },
-                        { "id": "network" },
-                        { "id": "mtu" }
+                    path: '/tm/net/route',
+                    schemaClass: 'Route',
+                    properties: [
+                        { id: 'gw' },
+                        { id: 'network' },
+                        { id: 'mtu' }
                     ]
                 }
             ];
@@ -228,14 +225,14 @@ describe('configManager', () => {
 
     it('should handle config items where we to map property name to value', () => {
         return new Promise((resolve, reject) => {
-            configItems = [
+            const configItems = [
                 {
-                    "path": "/tm/sys/provision",
-                    "schemaClass": "Provision",
-                    "properties": [
-                        { "id": "level" }
+                    path: '/tm/sys/provision',
+                    schemaClass: 'Provision',
+                    properties: [
+                        { id: 'level' }
                     ],
-                    "singleValue": true
+                    singleValue: true
                 }
             ];
 
@@ -255,7 +252,7 @@ describe('configManager', () => {
                 .then(() => {
                     assert.deepEqual(
                         state.currentConfig.Common.Provision,
-                        { 'afm': 'none', 'ltm': 'nominal' }
+                        { afm: 'none', ltm: 'nominal' }
                     );
                     resolve();
                 })
@@ -267,18 +264,18 @@ describe('configManager', () => {
 
     it('should handle references', () => {
         return new Promise((resolve, reject) => {
-            configItems = [
+            const configItems = [
                 {
-                    "path": "/tm/net/vlan",
-                    "schemaClass": "VLAN",
-                    "properties": [
-                        { "id": "mtu" },
-                        { "id": "tag" },
-                        { "id": "interfacesReference" }
+                    path: '/tm/net/vlan',
+                    schemaClass: 'VLAN',
+                    properties: [
+                        { id: 'mtu' },
+                        { id: 'tag' },
+                        { id: 'interfacesReference' }
                     ],
-                    "references": {
-                        "interfacesReference": [
-                            { "id": "tagged", "truth": true, "falsehood": false }
+                    references: {
+                        interfacesReference: [
+                            { id: 'tagged', truth: true, falsehood: false }
                         ]
                     }
                 }
@@ -318,12 +315,12 @@ describe('configManager', () => {
     describe('SelfIp oddities', () => {
         it('should strip /Common from vlan', () => {
             return new Promise((resolve, reject) => {
-                configItems = [
+                const configItems = [
                     {
-                        "path": "/tm/net/self",
-                        "schemaClass": "SelfIp",
-                        "properties": [
-                            { "id": "vlan" }
+                        path: '/tm/net/self',
+                        schemaClass: 'SelfIp',
+                        properties: [
+                            { id: 'vlan' }
                         ]
                     }
                 ];
@@ -349,12 +346,12 @@ describe('configManager', () => {
 
         it('should handle allowService with ["default"]', () => {
             return new Promise((resolve, reject) => {
-                configItems = [
+                const configItems = [
                     {
-                        "path": "/tm/net/self",
-                        "schemaClass": "SelfIp",
-                        "properties": [
-                            { "id": "allowService" }
+                        path: '/tm/net/self',
+                        schemaClass: 'SelfIp',
+                        properties: [
+                            { id: 'allowService' }
                         ]
                     }
                 ];
@@ -380,12 +377,12 @@ describe('configManager', () => {
 
         it('should handle allowService none', () => {
             return new Promise((resolve, reject) => {
-                configItems = [
+                const configItems = [
                     {
-                        "path": "/tm/net/self",
-                        "schemaClass": "SelfIp",
-                        "properties": [
-                            { "id": "allowService" }
+                        path: '/tm/net/self',
+                        schemaClass: 'SelfIp',
+                        properties: [
+                            { id: 'allowService' }
                         ]
                     }
                 ];
@@ -411,18 +408,18 @@ describe('configManager', () => {
 
     it('should set original config if missing', () => {
         return new Promise((resolve, reject) => {
-            configItems = [
+            const configItems = [
                 {
-                    "path": "/tm/sys/global-settings",
-                    "properties": [
-                        { "id": "hostname" }
+                    path: '/tm/sys/global-settings',
+                    properties: [
+                        { id: 'hostname' }
                     ]
                 },
                 {
-                    "path": "/tm/net/self",
-                    "schemaClass": "SelfIp",
-                    "properties": [
-                        { "id": "allowService" }
+                    path: '/tm/net/self',
+                    schemaClass: 'SelfIp',
+                    properties: [
+                        { id: 'allowService' }
                     ]
                 }
             ];
@@ -448,18 +445,18 @@ describe('configManager', () => {
 
     it('should not overwrite original config if present', () => {
         return new Promise((resolve, reject) => {
-            configItems = [
+            const configItems = [
                 {
-                    "path": "/tm/sys/global-settings",
-                    "properties": [
-                        { "id": "hostname" }
+                    path: '/tm/sys/global-settings',
+                    properties: [
+                        { id: 'hostname' }
                     ]
                 },
                 {
-                    "path": "/tm/net/self",
-                    "schemaClass": "SelfIp",
-                    "properties": [
-                        { "id": "allowService" }
+                    path: '/tm/net/self',
+                    schemaClass: 'SelfIp',
+                    properties: [
+                        { id: 'allowService' }
                     ]
                 }
             ];
@@ -498,14 +495,14 @@ describe('configManager', () => {
 
     describe('DbVariables', () => {
         it('should get DB variables if in declaration', () => {
-            configItems = [
+            const configItems = [
                 {
-                    "path": "/tm/sys/db",
-                    "schemaClass": "DbVariables",
-                    "properties": [
-                        { "id": "value" }
+                    path: '/tm/sys/db',
+                    schemaClass: 'DbVariables',
+                    properties: [
+                        { id: 'value' }
                     ],
-                    "singleValue": true
+                    singleValue: true
                 }
             ];
 
@@ -551,16 +548,16 @@ describe('configManager', () => {
 
     it('should ignore ignored properties', () => {
         return new Promise((resolve, reject) => {
-            configItems = [
+            const configItems = [
                 {
-                    "path": "/tm/cm/device-group",
-                    "schemaClass": "DeviceGroup",
-                    "properties": [
-                        { "id": "type" }
+                    path: '/tm/cm/device-group',
+                    schemaClass: 'DeviceGroup',
+                    properties: [
+                        { id: 'type' }
                     ],
-                    "ignore": [
-                        { "name": "^datasync-.+-dg$" },
-                        { "name": "^gtm$" }
+                    ignore: [
+                        { name: '^datasync-.+-dg$' },
+                        { name: '^gtm$' }
                     ]
                 }
             ];
@@ -594,5 +591,34 @@ describe('configManager', () => {
                     reject(err);
                 });
         });
+    });
+
+    it('should skip unprovisioned modules', () => {
+        const configItems = [
+            {
+                path: '/tm/analytics/global-settings',
+                requiredModule: 'avr',
+                schemaClass: 'Analytics',
+                properties: []
+            }
+        ];
+
+        const bigip = {
+            deviceInfo() {
+                return Promise.resolve({ hostname });
+            },
+            list(path) {
+                const pathname = URL.parse(path, 'https://foo').pathname;
+                assert(
+                    pathname !== '/tm/analytics/global-settings',
+                    'Analytics was checked but not provisioned'
+                );
+                return Promise.resolve(listResponses[pathname] || {});
+            }
+        };
+        const configManager = new ConfigManager(configItems, bigip);
+        return configManager.get({}, state, doState)
+            .then(() => {
+            });
     });
 });