+++ RELEASE_NOTES.md
@@ -7,6 +7,7 @@
     + BIG-IQ credentials not required in licensePool object when licensing from the BIG-IQ on which DO is running.
 + Fix bug in which credentials could appear in declaration results when revoking a license.
 + Disable DHCP for DNS/NTP if DO will be configuring them.
++ Allow setting global analytics settings.
 
 ## Version 1.4.1
 + Fix vulnerability CVE-2019-5021 in DO container